refactor: RecommendedRoutine 관련 공통 로직을 Service가 아닌 Manager 클래스로 위임하여 레이어드 아키텍처 규칙을 준수
